+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 ## Unreleased
 
 ### Added
-- Stable support for `prefix` and `suffix` query types.
+- Stable support for prefix and suffix queries:
+    - The `prefixPreview` query type is replaced with `prefix`.
+    - The `suffixPreview` query type is replaced with `suffix`.
+    - Use the `string` algorithm (formerly `textPreview`) for `prefix` and `suffix` query types.
 
 ### Removed
 - Support for the experimental `prefixPreview` and `suffixPreview` query types.